refactored a bit more
